Add vertical positioning for sections
Only the sections itself will be resized, so the footer is always at the
bottom. The elements inside the sections keep their size. Adjusting the
size of elements in relation to screen size is currently not necessary,
because it only comes into effect when using very big screen or zoom-out
level.